+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Decrease movespeed based on weight when using backpack.
+
 ## [0.4.0] - 2024-05-05
 
 ### Added